refactor : 변수명 변경 및 템플릿, 메뉴 갯수 count함수 utils로 분리
