--- src/js/utils/index.js
@@ -1,1 +1,0 @@
-export const $ = (selector) => document.querySelector(selector);
